automation/requirements.txt
sh==1.12.14
requests==2.24.0
python-box==4.0.2